+++ kyo-scheduler/jvm/src/main/scala/kyo/scheduler/regulator/Config.scala
@@ -1,0 +1,12 @@
+package kyo.scheduler.regulator
+
+case class Config(
+    collectWindowExp: Int,
+    collectIntervalMs: Int,
+    collectSamples: Int,
+    regulateIntervalMs: Int,
+    jitterUpperThreshold: Double,
+    jitterLowerThreshold: Double,
+    loadAvgTarget: Double,
+    stepExp: Double
+)
